iot-mqtt-to-kafka/requirements.txt
paho-mqtt==1.6.1
kafka-python==2.0.2
psycopg2-binary

